@@ -1,19 +1,19 @@
-﻿using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
+﻿using System.Text.Json;
+namespace FP.ContainerTraining.EventOperator.Business;
 
-namespace FP.ContainerTraining.EventOperator.Business
+public class KubernetesSerialization
 {
-    public class KubernetesSerialization
+    public static readonly JsonSerializerOptions Options = new JsonSerializerOptions
     {
-        public static readonly JsonSerializerSettings SerializerSettings = new JsonSerializerSettings
-        {
-            ContractResolver = new CamelCasePropertyNamesContractResolver()
-            {
-                NamingStrategy = new CamelCaseNamingStrategy
-                {
-                    ProcessDictionaryKeys = false
-                }
-            }
-        };
-    }
+        PropertyNamingPolicy = JsonNamingPolicy.CamelCase
+    };
+
+    // ContractResolver = new CamelCasePropertyNamesContractResolver()
+    // {
+    //     NamingStrategy = new CamelCaseNamingStrategy
+    //     {
+    //         ProcessDictionaryKeys = false
+    //     }
+    // }
+
 }
